requirements.txt
streamlit==1.51.0
streamlit-autorefresh==1.0.1
fastapi==0.121.1
uvicorn==0.38.0
python-dotenv==1.2.1
requests==2.32.5
pydantic==2.12.4
pydantic_core==2.41.5
SQLAlchemy>=1.4
psycopg2-binary
typing_extensions
